refactor: Spring Framework 의존성 제거
