+++ jdbc/src/main/java/nextstep/jdbc/exception/IncorrectResultSizeDataAccessException.java
@@ -1,0 +1,8 @@
+package nextstep.jdbc.exception;
+
+public class IncorrectResultSizeDataAccessException extends DataAccessException {
+
+    public IncorrectResultSizeDataAccessException(String message) {
+        super(message);
+    }
+}
